% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scspatstat.r
\name{sc_coord}
\alias{sc_coord}
\alias{sc_coord.owin}
\alias{sc_coord}
\alias{sc_coord.ppp}
\alias{sc_coord}
\alias{sc_coord.psp}
\title{Coord}
\usage{
\method{sc_coord}{owin}(x, ...)

\method{sc_coord}{ppp}(x, ...)

\method{sc_coord}{psp}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
sc coord
}
\examples{
irreg_data <- c("chorley", "clmfires", "demopat", "gordon", "gorillas", "humberside",
"murchison", "nbfires", "urkiola", "vesicles")
## all are ppp, but murchison  "solist"  "anylist" "listof"  "list"
library(spatstat)
data(list = irreg_data, package = "spatstat")
for (i in seq_along(irreg_data)) {
x <- get(irreg_data[i])
 #plot(x, main = irreg_data[i])
 # title(sub = paste(class(x), collapse = " "))
   if (inherits(x, "solist")) {
   next;
  }
   #print(sc_coord(as.owin(x)))
   #print(sc_path(as.owin(x)))
   #print(sc_object(as.owin(x)))
   print(sc_object(x))
}
x <- as.owin(x)
sc_path(x)
}
