% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scspatstat.r
\name{sc_path}
\alias{sc_path}
\alias{sc_path.owin}
\alias{sc_path}
\alias{sc_path.ppp}
\alias{sc_path}
\alias{sc_path.psp}
\title{Path}
\usage{
\method{sc_path}{owin}(x, ids = NULL, ...)

\method{sc_path}{ppp}(x, ids = NULL, ...)

\method{sc_path}{psp}(x, ids = NULL, ...)
}
\arguments{
\item{x}{input object}

\item{...}{arguments passed to methods}
}
\description{
sc path
}
